Simplified code thanks to C++14 variable templates
