+++ include/calculate/util.hpp
@@ -1,6 +1,6 @@
 /*
     Calculate - Version 2.1.1dev0
-    Last modified 2018/05/18
+    Last modified 2018/06/03
     Released under MIT license
     Copyright (c) 2016-2018 Alberto Lorenzo <alorenzo.md@gmail.com>
 */
@@ -12,7 +12,8 @@
 #include <functional>
 #include <memory>
 #include <regex>
-#include <unordered_map>
+#include <tuple>
+#include <type_traits>
 #include <vector>
 
 #include "exception.hpp"
@@ -21,6 +22,22 @@
 namespace calculate {
 
 namespace util {
+
+template<typename T, typename U>
+constexpr bool is_same = std::is_same<T, U>::value;
+
+template<typename T, typename U>
+constexpr bool is_base_of = std::is_base_of<T, U>::value;
+
+template<typename T>
+constexpr bool is_integral = std::is_integral<T>::value;
+
+template<typename T>
+constexpr bool is_copy_constructible = std::is_copy_constructible<T>::value;
+
+template<typename T>
+constexpr std::size_t tuple_size = std::tuple_size<T>::value;
+
 
 namespace detail {
 
@@ -38,12 +55,108 @@ static constexpr decltype(
 template<typename Type>
 static constexpr bool is_iterable(...) { return false; }
 
+
+template<typename Function, typename... Args>
+constexpr bool is_noexcept =
+    noexcept(std::declval<Function>()(std::declval<Args>()...));
+
+template<typename Type, typename = void>
+struct Traits : Traits<decltype(&Type::operator())> {};
+
+template<typename Result, typename... Args>
+struct Traits<std::function<Result(Args...)>, void> {
+    using result = Result;
+    using arguments = std::tuple<std::decay_t<Args>...>;
+    static constexpr bool constant = true;
+};
+
+template<typename Result, typename... Args>
+struct Traits<
+    Result(*)(Args...) noexcept,
+    std::enable_if_t<is_noexcept<Result(*)(Args...) noexcept, Args...>>
+> {
+    using result = Result;
+    using arguments = std::tuple<std::decay_t<Args>...>;
+    static constexpr bool constant = true;
+};
+
+template<typename Result, typename... Args>
+struct Traits<
+    Result(*)(Args...),
+    std::enable_if_t<!is_noexcept<Result(*)(Args...), Args...>>
+> {
+    using result = Result;
+    using arguments = std::tuple<std::decay_t<Args>...>;
+    static constexpr bool constant = true;
+};
+
+template<typename Type, typename Result, typename... Args>
+struct Traits<
+    Result(Type::*)(Args...) noexcept,
+    std::enable_if_t<is_noexcept<Type, Args...>>
+> {
+    using result = Result;
+    using arguments = std::tuple<std::decay_t<Args>...>;
+    static constexpr bool constant = false;
+};
+
+template<typename Type, typename Result, typename... Args>
+struct Traits<
+    Result(Type::*)(Args...),
+    std::enable_if_t<!is_noexcept<Type, Args...>>
+> {
+    using result = Result;
+    using arguments = std::tuple<std::decay_t<Args>...>;
+    static constexpr bool constant = false;
+};
+
+template<typename Type, typename Result, typename... Args>
+struct Traits<
+    Result(Type::*)(Args...) const noexcept,
+    std::enable_if_t<is_noexcept<Type, Args...>>
+> {
+    using result = Result;
+    using arguments = std::tuple<std::decay_t<Args>...>;
+    static constexpr bool constant = true;
+};
+
+template<typename Type, typename Result, typename... Args>
+struct Traits<
+    Result(Type::*)(Args...) const,
+    std::enable_if_t<!is_noexcept<Type, Args...>>
+> {
+    using result = Result;
+    using arguments = std::tuple<std::decay_t<Args>...>;
+    static constexpr bool constant = true;
+};
+
 }
 
 template<typename Type>
-struct Check {
-    static constexpr bool iterable = detail::is_iterable<Type>(0);
-};
+constexpr bool is_iterable = detail::is_iterable<Type>(0);
+
+template<typename Type>
+constexpr bool is_noexcept = detail::is_noexcept<Type>;
+
+
+template<typename Function>
+using Result = typename detail::Traits<Function>::result;
+
+template<typename Function>
+constexpr bool is_const = detail::Traits<Function>::constant;
+
+template<typename Function>
+using Arguments = typename detail::Traits<Function>::arguments;
+
+template<typename Function>
+constexpr std::size_t argc =
+    util::tuple_size<typename detail::Traits<Function>::arguments>;
+
+template<typename Type, typename Target>
+constexpr bool not_same =
+    !is_same<std::decay_t<Type>, Target> &&
+    !is_base_of<Target, std::decay_t<Type>>;
+
 
 template<typename Type>
 const std::vector<Type>& to_vector(const std::vector<Type>& args) { return args; }
@@ -52,7 +165,7 @@ template<typename Type, typename... Args>
 std::vector<Type> to_vector(Args&&... args) { return {std::forward<Args>(args)...}; }
 
 template<typename Type, typename Args>
-std::enable_if_t<Check<Args>::iterable, std::vector<Type>>
+std::enable_if_t<is_iterable<Args>, std::vector<Type>>
 to_vector(Args&& args) { return {std::begin(args), std::end(args)}; }
 
 
@@ -63,199 +176,6 @@ void hash_combine(std::size_t& seed, const Type& object) {
 }
 
 }
-
-
-template<typename Kind, typename Parser>
-class SymbolContainer final : std::unordered_map<std::string, Kind> {
-    friend Parser;
-    using Base = std::unordered_map<std::string, Kind>;
-    using Lexer = typename Parser::Lexer;
-
-    using Constant = typename Parser::Constant;
-    using Function = typename Parser::Function;
-    using Operator = typename Parser::Operator;
-
-public:
-    using typename Base::key_type;
-    using typename Base::mapped_type;
-    using typename Base::value_type;
-    using typename Base::iterator;
-    using typename Base::const_iterator;
-
-private:
-    Lexer* _lexer;
-
-    SymbolContainer(Lexer* lexer) : _lexer{lexer} {}
-
-    SymbolContainer(const SymbolContainer&) = default;
-    SymbolContainer(SymbolContainer&&) = default;
-    ~SymbolContainer() = default;
-
-    SymbolContainer& operator=(const SymbolContainer&) = default;
-    SymbolContainer& operator=(SymbolContainer&&) = default;
-
-    void _validate(const std::string& key, Constant*) const {
-        if (!std::regex_match(key, _lexer->name_regex))
-            throw UnsuitableName{key};
-    }
-
-    void _validate(const std::string& key, Function*) const {
-        if (!std::regex_match(key, _lexer->name_regex))
-            throw UnsuitableName{key};
-    }
-
-    void _validate(const std::string& key, Operator*) const {
-        if (!std::regex_match(key, _lexer->symbol_regex))
-            throw UnsuitableName{key};
-    }
-
-    void _validate(const std::string& key) const {
-        _validate(key, static_cast<mapped_type*>(nullptr));
-    }
-
-public:
-    using Base::begin;
-    using Base::end;
-    using Base::cbegin;
-    using Base::cend;
-
-    using Base::empty;
-    using Base::size;
-    using Base::find;
-    using Base::count;
-    using Base::at;
-
-    using Base::erase;
-    using Base::clear;
-    using Base::swap;
-    using Base::reserve;
-
-    mapped_type& operator[](const key_type& key) {
-        return Base::find(key)->second;
-    }
-
-    mapped_type& operator[](key_type&& key) {
-        return Base::find(std::move(key))->second;
-    }
-
-    template<typename... Args>
-    std::pair<iterator, bool> emplace(const std::string& key, Args&&... args) {
-        _validate(key);
-        return Base::emplace(key, std::forward<Args>(args)...);
-    }
-
-    std::pair<iterator, bool> insert(const value_type& value) {
-        _validate(value.first);
-        return Base::insert(value);
-    }
-
-    template<typename Value>
-    std::pair<iterator, bool> insert(Value&& value) {
-        _validate(value.first);
-        return Base::insert(std::forward<Value>(value));
-    }
-
-    template<typename Iterator>
-    void insert(Iterator first, Iterator last) {
-        for (auto element = first; element != last; ++element)
-            _validate(element->first);
-        Base::insert(first, last);
-    }
-
-    void insert(std::initializer_list<value_type> list) {
-        for (const auto& element : list)
-            _validate(element.first);
-        Base::insert(list);
-    }
-};
-
-template<typename Parser>
-class AliasContainer final : std::unordered_map<std::string, std::string> {
-    friend Parser;
-    using Base = std::unordered_map<std::string, std::string>;
-    using Lexer = typename Parser::Lexer;
-
-public:
-    using typename Base::key_type;
-    using typename Base::mapped_type;
-    using typename Base::value_type;
-    using typename Base::iterator;
-    using typename Base::const_iterator;
-
-private:
-    Lexer* _lexer;
-
-    AliasContainer(Lexer* lexer) : _lexer{lexer} {}
-
-    AliasContainer(const AliasContainer&) = default;
-    AliasContainer(AliasContainer&&) = default;
-    ~AliasContainer() = default;
-
-    AliasContainer& operator=(const AliasContainer&) = default;
-    AliasContainer& operator=(AliasContainer&&) = default;
-
-    void _validate(const std::string& key, const std::string& value) const {
-        if (!std::regex_match(key, _lexer->symbol_regex))
-            throw UnsuitableName{key};
-        if (!std::regex_match(value, _lexer->name_regex))
-            throw UnsuitableName{value};
-    }
-
-public:
-    using Base::begin;
-    using Base::end;
-    using Base::cbegin;
-    using Base::cend;
-
-    using Base::empty;
-    using Base::size;
-    using Base::find;
-    using Base::count;
-    using Base::at;
-
-    using Base::erase;
-    using Base::clear;
-    using Base::swap;
-    using Base::reserve;
-
-    mapped_type& operator[](const key_type& key) {
-        return Base::find(key)->second;
-    }
-
-    mapped_type& operator[](key_type&& key) {
-        return Base::find(std::move(key))->second;
-    }
-
-    template<typename Value>
-    std::pair<iterator, bool> emplace(const std::string& key, Value&& value) {
-        _validate(key, value);
-        return Base::emplace(key, std::forward<Value>(value));
-    }
-
-    std::pair<iterator, bool> insert(const value_type& value) {
-        _validate(value.first, value.second);
-        return Base::insert(value);
-    }
-
-    template<typename Value>
-    std::pair<iterator, bool> insert(Value&& value) {
-        _validate(value.first, value.second);
-        return Base::insert(std::forward<Value>(value));
-    }
-
-    template<typename Iterator>
-    void insert(Iterator first, Iterator last) {
-        for (auto element = first; element != last; ++element)
-            _validate(element->first, element->second);
-        Base::insert(first, last);
-    }
-
-    void insert(std::initializer_list<value_type> list) {
-        for (const auto& element : list)
-            _validate(element.first, element.second);
-        Base::insert(list);
-    }
-};
 
 }
